pyproject.toml
[project]
name = "sdlc-assist-mcp"
version = "0.1.0"
description = "MCP server providing access to SDLC Assist project artifacts (Supabase) and agent-delegated tools (Vertex AI)"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "mcp>=1.9.0",
    "httpx>=0.27.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "google-auth>=2.0.0",
    "requests>=2.20.0",
]

[project.scripts]
sdlc-assist-mcp = "sdlc_assist_mcp.server:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/sdlc_assist_mcp"]
